66 Wilson Gardens
"'The sundial in the garden. There is no other,' said I; 'but the papers must be those that are
destroyed.'

"'Pooh!' said he, gripping hard at his courage. 'We are in a civilised land here, and we can't have
tomfoolery of this kind. Where does the thing come from?'

"'From Dundee,' I answered, glancing at the postmark.

"'Some preposterous practical joke,' said he. 'What have I to do with sundials and papers? I shall
take no notice of such nonsense.'

"'I should certainly speak to the police,' I said.

"'And be laughed at for my pains. Nothing of the sort.'