13 Down Close
We were as good as our word, for it was just seven when we reached the Copper Beeches, having put up
our trap at a wayside public-house. The group of trees, with their dark leaves shining like
burnished metal in the light of the setting sun, were sufficient to mark the house even had Miss
Hunter not been standing smiling on the door-step.

"Have you managed it?" asked Holmes.