34 Tamworth Drive
"I have very little difficulty in finding what I want," said I, "for the facts are quite recent, and
the matter struck me as remarkable. I feared to refer them to you, however, as I knew that you had
an inquiry on hand and that you disliked the intrusion of other matters."

"Oh, you mean the little problem of the Grosvenor Square furniture van. That is quite cleared up now
--though, indeed, it was obvious from the first. Pray give me the results of your newspaper
selections."

"Here is the first notice which I can find. It is in the personal column of the Morning Post, and
dates, as you see, some weeks back:

"'A marriage has been arranged [it says] and will, if rumour is correct, very shortly take place,
between Lord Robert St. Simon, second son of the Duke of Balmoral, and Miss Hatty Doran, the only
daughter of Aloysius Doran. Esq., of San Francisco, Cal., U.S.A.'