30 Wilson Gardens
"The further points, that he is middle-aged, that his hair is grizzled, that it has been recently
cut, and that he uses lime-cream, are all to be gathered from a close examination of the lower part
of the lining. The lens discloses a large number of hair-ends, clean cut by the scissors of the
barber. They all appear to be adhesive, and there is a distinct odour of lime-cream. This dust, you
will observe, is not the gritty, grey dust of the street but the fluffy brown dust of the house,
showing that it has been hung up indoors most of the time, while the marks of moisture upon the
inside are proof positive that the wearer perspired very freely, and could therefore, hardly be in
the best of training."

"But his wife--you said that she had ceased to love him."

"This hat has not been brushed for weeks. When I see you, my dear Watson, with a week's accumulation
of dust upon your hat, and when your wife allows you to go out in such a state, I shall fear that
you also have been unfortunate enough to lose your wife's affection."

"But he might be a bachelor."