101 Cliff Place
Our visitor collapsed into a chair, with a ghastly face and a glitter of moisture on his brow. "It--
it's not actionable," he stammered.

"I am very much afraid that it is not. But between ourselves, Windibank, it was as cruel and selfish
and heartless a trick in a petty way as ever came before me. Now, let me just run over the course of
events, and you will contradict me if I go wrong."

The man sat huddled up in his chair, with his head sunk upon his breast, like one who is utterly
crushed. Holmes stuck his feet up on the corner of the mantelpiece and, leaning back with his hands
in his pockets, began talking, rather to himself, as it seemed, than to us.