102 Larch Walk
"The grass was growing under it. It had only lain there a few days. There was no sign of a place
whence it had been taken. It corresponds with the injuries. There is no sign of any other weapon."

"And the murderer?"

"Is a tall man, left-handed, limps with the right leg, wears thick-soled shooting-boots and a grey
cloak, smokes Indian cigars, uses a cigar-holder, and carries a blunt pen-knife in his pocket. There
are several other indications, but these may be enough to aid us in our search."

Lestrade laughed. "I am afraid that I am still a sceptic," he said. "Theories are all very well, but
we have to deal with a hard-headed British jury."

"Nous verrons," answered Holmes calmly. "You work your own method, and I shall work mine. I shall be
busy this afternoon, and shall probably return to London by the evening train."