102 Sunningdale Drive
"It is a curious thing," remarked Holmes, "that a typewriter has really quite as much individuality
as a man's handwriting. Unless they are quite new, no two of them write exactly alike. Some letters
get more worn than others, and some wear only on one side. Now, you remark in this note of yours,
Mr. Windibank, that in every case there is some little slurring over of the 'e,' and a slight defect
in the tail of the 'r.' There are fourteen other characteristics, but those are the more obvious."

"We do all our correspondence with this machine at the office, and no doubt it is a little worn,"
our visitor answered, glancing keenly at Holmes with his bright little eyes.

"And now I will show you what is really a very interesting study, Mr. Windibank," Holmes continued.
"I think of writing another little monograph some of these days on the typewriter and its relation
to crime. It is a subject to which I have devoted some little attention. I have here four letters
which purport to come from the missing man. They are all typewritten. In each case, not only are the
'e's' slurred and the 'r's' tailless, but you will observe, if you care to use my magnifying lens,
that the fourteen other characteristics to which I have alluded are there as well."

Mr. Windibank sprang out of his chair and picked up his hat. "I cannot waste time over this sort of
fantastic talk, Mr. Holmes," he said. "If you can catch the man, catch him, and let me know when you
have done it."