74 Tamworth Drive
"What can it mean?" I gasped.

"It means that it is all over," Holmes answered. "And perhaps, after all, it is for the best. Take
your pistol, and we will enter Dr. Roylott's room."

With a grave face he lit the lamp and led the way down the corridor. Twice he struck at the chamber
door without any reply from within. Then he turned the handle and entered, I at his heels, with the
cocked pistol in my hand.