21 Tamworth Drive
"I am glad of all details," remarked my friend, "whether they seem to you to be relevant or not."

"I shall try not to miss anything of importance. The one unpleasant thing about the house, which
struck me at once, was the appearance and conduct of the servants. There are only two, a man and his
wife. Toller, for that is his name, is a rough, uncouth man, with grizzled hair and whiskers, and a
perpetual smell of drink. Twice since I have been with them he has been quite drunk, and yet Mr.
Rucastle seemed to take no notice of it. His wife is a very tall and strong woman with a sour face,
as silent as Mrs. Rucastle and much less amiable. They are a most unpleasant couple, but fortunately
I spend most of my time in the nursery and my own room, which are next to each other in one corner
of the building.

"For two days after my arrival at the Copper Beeches my life was very quiet; on the third, Mrs.
Rucastle came down just after breakfast and whispered something to her husband.