150 Sunningdale Drive
"In that case I shall continue my professional round. But I shall come back in the evening at the
hour you have mentioned, for I should like to see the solution of so tangled a business."

"Very glad to see you. I dine at seven. There is a woodcock, I believe. By the way, in view of
recent occurrences, perhaps I ought to ask Mrs. Hudson to examine its crop."

I had been delayed at a case, and it was a little after half-past six when I found myself in Baker
Street once more. As I approached the house I saw a tall man in a Scotch bonnet with a coat which
was buttoned up to his chin waiting outside in the bright semicircle which was thrown from the
fanlight. Just as I arrived the door was opened, and we were shown up together to Holmes' room.

"Mr. Henry Baker, I believe," said he, rising from his armchair and greeting his visitor with the
easy air of geniality which he could so readily assume. "Pray take this chair by the fire, Mr.
Baker. It is a cold night, and I observe that your circulation is more adapted for summer than for
winter. Ah, Watson, you have just come at the right time. Is that your hat, Mr. Baker?"

"Yes, sir, that is undoubtedly my hat."