32 Tamworth Drive
"I have said that though the floor and ceiling were of iron, the walls were of wood. As I gave a
last hurried glance around, I saw a thin line of yellow light between two of the boards, which
broadened and broadened as a small panel was pushed backward. For an instant I could hardly believe
that here was indeed a door which led away from death. The next instant I threw myself through, and
lay half-fainting upon the other side. The panel had closed again behind me, but the crash of the
lamp, and a few moments afterwards the clang of the two slabs of metal, told me how narrow had been
my escape.

"I was recalled to myself by a frantic plucking at my wrist, and I found myself lying upon the stone
floor of a narrow corridor, while a woman bent over me and tugged at me with her left hand, while
she held a candle in her right. It was the same good friend whose warning I had so foolishly
rejected.

"'Come! come!' she cried breathlessly. 'They will be here in a moment. They will see that you are
not there. Oh, do not waste the so-precious time, but come!'

"This time, at least, I did not scorn her advice. I staggered to my feet and ran with her along the
corridor and down a winding stair. The latter led to another broad passage, and just as we reached
it we heard the sound of running feet and the shouting of two voices, one answering the other from
the floor on which  we were and from the one beneath. My guide stopped and looked about her like one
who is at her wit's end. Then she threw open a door which led into a bedroom, through the window of
which the moon was shining brightly.

"'It is your only chance,' said she. 'It is high, but it may be that you can jump it.'

"As she spoke a light sprang into view at the further end of the passage, and I saw the lean figure
of Colonel Lysander Stark rushing forward with a lantern in one hand and a weapon like a butcher's
cleaver in the other. I rushed across the bedroom, flung open the window, and looked out. How quiet
and sweet and wholesome the garden looked in the moonlight, and it could not be more than thirty
feet down. I clambered out upon the sill, but I hesitated to jump until I should have heard what
passed between my saviour and the ruffian who pursued me. If she were ill-used, then at any risks I
was determined to go back to her assistance. The thought had hardly flashed through my mind before
he was at the door, pushing his way past her; but she threw her arms round him and tried to hold him
back.