98 Glan Road
"Now, my dear sir," said Holmes. "is it not obvious to you now that this matter really strikes very
much deeper than either you or the police were at first inclined to think? It appeared to you to be
a simple case; to me it seems exceedingly complex. Consider what is involved by your theory. You
suppose that your son came down from his bed, went, at great risk, to your dressing-room, opened
your bureau, took out your coronet, broke off by main force a small portion of it, went off to some
other place, concealed three gems out of the thirty-nine, with such skill that nobody can find them,
and then returned with the other thirty-six into the room in which he exposed himself to the
greatest danger of being discovered. I ask you now, is such a theory tenable?"

"But what other is there?" cried the banker with a gesture of despair. "If his motives were
innocent, why does he not explain them?"

"It is our task to find that out," replied Holmes; "so now, if you please, Mr. Holder, we will set
off for Streatham together, and devote an hour to glancing a little more closely into details."

My friend insisted upon my accompanying them in their expedition, which I was eager enough to do,
for my curiosity and sympathy were deeply stirred by the story to which we had listened. I confess
that the guilt of the banker's son appeared to me to be as obvious as it did to his unhappy father,
but still I had such faith in Holmes' judgment that I felt that there must be some grounds for hope
as long as he was dissatisfied with the accepted explanation. He hardly spoke a word the whole way
out to the southern suburb, but sat with his chin upon his breast and his hat drawn over his eyes,
sunk in the deepest thought. Our client appeared to have taken fresh heart at the little glimpse of
hope which had been presented to him, and he even broke into a desultory chat with me over his
business affairs. A short railway journey and a shorter walk brought us to Fairbank, the modest
residence of the great financier.

Fairbank was a good-sized square house of white stone, standing back a little from the road. A
double carriage-sweep, with a snow-clad lawn, stretched down in front to two large iron gates which
closed the entrance. On the right side was a small wooden thicket, which led into a narrow path
between two neat hedges stretching from the road to the kitchen door, and forming the tradesmen's
entrance. On the left ran a lane which led to the stables, and was not itself within the grounds at
all, being a public, though little used, thoroughfare. Holmes left us standing at the door and
walked slowly all round the house, across the front, down the tradesmen's path, and so round by the
garden behind into the stable lane. So long was he that Mr. Holder and I went into the dining-room
and waited by the fire until he should return. We were sitting there in silence when the door opened
and a young lady came in. She was rather above the middle height, slim, with dark hair and eyes,
which seemed the darker against the absolute pallor of her skin. I do not think that I have ever
seen such deadly paleness in a woman's face. Her lips, too, were bloodless, but her eyes were
flushed with crying. As she swept silently into the room she impressed me with a greater sense of
grief than the banker had done in the morning, and it was the more striking in her as she was
evidently a woman of strong character, with immense capacity for self-restraint. Disregarding my
presence, she went straight to her uncle and passed her hand over his head with a sweet womanly
caress.

"You have given orders that Arthur should be liberated, have you not, dad?" she asked.