29 Glan Road
"It has been in some points a singular case," said Holmes, flicking the horse on into a gallop. "I
confess that I have been as blind as a mole, but it is better to learn wisdom late than never to
learn it at all."

In town the earliest risers were just beginning to look sleepily from their windows as we drove
through the streets of the Surrey side. Passing down the Waterloo Bridge Road we crossed over the
river, and dashing up Wellington Street wheeled sharply to the right and found ourselves in Bow
Street. Sherlock Holmes was well known to the force, and the two constables at the door saluted him.
One of them held the horse's head while the other led us in.

"Who is on duty?" asked Holmes.

"Inspector Bradstreet, sir."

"Ah, Bradstreet, how are you?" A tall, stout official had come down the stone-flagged passage, in a
peaked cap and frogged jacket. "I wish to have a quiet word with you, Bradstreet." "Certainly, Mr.
Holmes. Step into my room here." It was a small, office-like room, with a huge ledger upon the
table, and a telephone projecting from the wall. The inspector sat down at his desk.

"What can I do for you, Mr. Holmes?"