45 Valentia Road
"A pair, by the sound," said he. "Yes," he continued, glancing out of the window. "A nice little
brougham and a pair of beauties. A hundred and fifty guineas apiece. There's money in this case,
Watson, if there is nothing else."

"I think that I had better go, Holmes."

"Not a bit, Doctor. Stay where you are. I am lost without my Boswell. And this promises to be
interesting. It would be a pity to miss it."

"But your client--"