55 Harvey Street
"That is a possible solution--in fact, as matters stand, it is the most probable one. But in any
case it does not seem to be a nice household for a young lady."

"But the money, Mr. Holmes, the money!"

"Well, yes, of course the pay is good--too good. That is what makes me uneasy. Why should they give
you £120 a year, when they could have their pick for £40? There must be some strong reason behind."

"I thought that if I told you the circumstances you would understand afterwards if I wanted your
help. I should feel so much stronger if I felt that you were at the back of me."