63 Tamworth Drive
"Oh, yes, Mr. Holmes. We were engaged after the first walk that we took. Hosmer--Mr. Angel--was a
cashier in an office in Leadenhall Street--and--"

"What office?"

"That's the worst of it, Mr. Holmes, I don't know."

"Where did he live, then?"

"He slept on the premises."

"And you don't know his address?"