21 Valentia Road
"We're hunting in couples again, Doctor, you see," said Jones in his consequential way. "Our friend
here is a wonderful man for starting a chase. All he wants is an old dog to help him to do the
running down."

"I hope a wild goose may not prove to be the end of our chase," observed Mr. Merryweather gloomily.

"You may place considerable confidence in Mr. Holmes, sir," said the police agent loftily. "He has
his own little methods, which are, if he won't mind my saying so, just a little too theoretical and
fantastic, but he has the makings of a detective in him. It is not too much to say that once or
twice, as in that business of the Sholto murder and the Agra treasure, he has been more nearly
correct than the official force."

"Oh, if you say so, Mr. Jones, it is all right," said the stranger with deference. "Still, I confess
that I miss my rubber. It is the first Saturday night for seven-and-twenty years that I have not had
my rubber."