80 Larch Walk
"Well, obviously it could not have been meant for the son. The son, as far as he knew, was in
Bristol. It was mere chance that he was within earshot. The 'Cooee!' was meant to attract the
attention of whoever it was that he had the appointment with. But 'Cooee' is a distinctly Australian
cry, and one which is used between Australians. There is a strong presumption that the person whom
McCarthy expected to meet him at Boscombe Pool was someone who had been in Australia."

"What of the rat, then?"

Sherlock Holmes took a folded paper from his pocket and flattened it out on the table. "This is a
map of the Colony of Victoria," he said. "I wired to Bristol for it last night." He put his hand
over part of the map. "What do you read?"

"ARAT," I read.

"And now?" He raised his hand.