124 Larch Walk
"We have got to the deductions and the inferences," said Lestrade, winking at me. "I find it hard
enough to tackle facts, Holmes, without flying away after theories and fancies."

"You are right," said Holmes demurely; "you do find it very hard to tackle the facts."

"Anyhow, I have grasped one fact which you seem to find it difficult to get hold of," replied
Lestrade with some warmth.

"And that is--"

"That McCarthy senior met his death from McCarthy junior and that all theories to the contrary are
the merest moonshine."