123 Badgers Dene
"What is the meaning of it all, Mr. Holmes?"

"Ah, I have no data. I cannot tell. Perhaps you have yourself formed some opinion?"

"Well, there seems to me to be only one possible solution. Mr. Rucastle seemed to be a very kind,
good-natured man. Is it not possible that his wife is a lunatic, that he desires to keep the matter
quiet for fear she should be taken to an asylum, and that he humours her fancies in every way in
order to prevent an outbreak?"