168 Beaconside
"Pray take a seat," said Holmes. "This is my friend and colleague, Dr. Watson, who is occasionally
good enough to help me in my cases. Whom have I the honour to address?"

"You may address me as the Count Von Kramm, a Bohemian nobleman. I understand that this gentleman,
your friend, is a man of honour and discretion, whom I may trust with a matter of the most extreme
importance. If not, I should much prefer to communicate with you alone."

I rose to go, but Holmes caught me by the wrist and pushed me back into my chair. "It is both, or
none," said he. "You may say before this gentleman anything which you may say to me."