35 Wellswood Gardens
"You have done well indeed!" cried Holmes with enthusiasm. "Now lead the way, and we shall soon see
the end of this black business."

We passed up the stair, unlocked the door, followed on down a passage, and found ourselves in front
of the barricade which Miss Hunter had described. Holmes cut the cord and removed the transverse
bar. Then he tried the various keys in the lock, but without success. No sound came from within, and
at the silence Holmes' face clouded over.

"I trust that we are not too late," said he. "I think, Miss Hunter, that we had better go in without
you. Now, Watson, put your shoulder to it, and we shall see whether we cannot make our way in."