75 Moor End
"There was a double line of tracks of a booted man, and a second double line which I saw with
delight belonged to a man with naked feet. I was at once convinced from what you had told me that
the latter was your son. The first had walked both ways, but the other had run swiftly, and as his
tread was marked in places over the depression of the boot, it was obvious that he had passed after
the other. I followed them up and found they led to the hall window, where Boots had worn all the
snow away while waiting. Then I walked to the other end, which was a hundred yards or more down the
lane. I saw where Boots had faced round, where the snow was cut up as though there had been a
struggle, and, finally, where a few drops of blood had fallen, to show me that I was not mistaken.
Boots had then run down the lane, and another little smudge of blood showed that it was he who had
been hurt. When he came to the highroad at the other end, I found that the pavement had been
cleared, so there was an end to that clue.

"On entering the house, however, I examined, as you remember, the sill and framework of the hall
window with my lens, and I could at once see that someone had passed out. I could distinguish the
outline of an instep where the wet foot had been placed in coming in. I was then beginning to be
able to form an opinion as to what had occurred. A man had waited outside the window; someone had
brought the gems; the deed had been overseen by your son; he had pursued the thief; had struggled
with him; they had each tugged at the coronet, their united strength causing injuries which neither
alone could have effected. He had returned with the prize, but had left a fragment in the grasp of
his opponent. So far I was clear. The question now was, who was the man and who was it brought him
the coronet?

"It is an old maxim of mine that when you have excluded the impossible, whatever remains, however
improbable, must be the truth. Now, I knew that it was not you who had brought it down, so there
only remained your niece and the maids. But if it were the maids, why should your son allow himself
to be accused in their place? There could be no possible reason. As he loved his cousin, however,
there was an excellent explanation why he should retain her secret--the more so as the secret was a
disgraceful one. When I remembered that you had seen her at that window, and how she had fainted on
seeing the coronet again, my conjecture became a certainty.