82 Cliff Place
A man entered who could hardly have been less than six feet six inches in height, with the chest and
limbs of a Hercules. His dress was rich with a richness which would, in England, be looked upon as
akin to bad taste. Heavy bands of astrakhan were slashed across the sleeves and fronts of his
double-breasted coat, while the deep blue cloak which was thrown over his shoulders was lined with
flame-coloured silk and secured at the neck with a brooch which consisted of a single flaming beryl.
Boots which extended halfway up his calves, and which were trimmed at the tops with rich brown fur,
completed the impression of barbaric opulence which was suggested by his whole appearance. He
carried a broad-brimmed hat in his hand, while he wore across the upper part of his face, extending
down past the cheekbones, a black vizard mask, which he had apparently adjusted that very moment,
for his hand was still raised to it as he entered. From the lower part of the face he appeared to be
a man of strong character, with a thick, hanging lip, and a long, straight chin suggestive of
resolution pushed to the length of obstinacy.

"You had my note?" he asked with a deep harsh voice and a strongly marked German accent. "I told you
that I would call." He looked from one to the other of us, as if uncertain which to address.

"Pray take a seat," said Holmes. "This is my friend and colleague, Dr. Watson, who is occasionally
good enough to help me in my cases. Whom have I the honour to address?"

"You may address me as the Count Von Kramm, a Bohemian nobleman. I understand that this gentleman,
your friend, is a man of honour and discretion, whom I may trust with a matter of the most extreme
importance. If not, I should much prefer to communicate with you alone."

I rose to go, but Holmes caught me by the wrist and pushed me back into my chair. "It is both, or
none," said he. "You may say before this gentleman anything which you may say to me."

The Count shrugged his broad shoulders. "Then I must begin," said he, "by binding you both to
absolute secrecy for two years; at the end of that time the matter will be of no importance. At
present it is not too much to say that it is of such weight it may have an influence upon European
history."