99 Sunningdale Drive
"'Or to sit here, or sit there, that would not be offensive to you?'

"'Oh, no.'

"'Or to cut your hair quite short before you come to us?'

"I could hardly believe my ears. As you may observe, Mr. Holmes, my hair is somewhat luxuriant, and
of a rather peculiar tint of chestnut. It has been considered artistic. I could not dream of
sacrificing it in this offhand fashion.