59 Badgers Dene
"Oh, yes; plenty. Then there is another note in the Morning Post to say that the marriage would be
an absolutely quiet one, that it would be at St. George's, Hanover Square, that only half a dozen
intimate friends would be invited, and that the party would return to the furnished house at
Lancaster Gate which has been taken by Mr. Aloysius Doran. Two days later--that is, on Wednesday
last--there is a curt announcement that the wedding had taken place, and that the honeymoon would be
passed at Lord Backwater's place, near Petersfield. Those are all the notices which appeared before
the disappearance of the bride."

"Before the what?" asked Holmes with a start.

"The vanishing of the lady."

"When did she vanish, then?"

"At the wedding breakfast."