7 Larch Walk
"Get back into your chair!" said Holmes sternly. "It is very well to cringe and crawl now, but you
thought little enough of this poor Horner in the dock for a crime of which he knew nothing."

"I will fly, Mr. Holmes. I will leave the country, sir. Then the charge against him will break
down."

"Hum! We will talk about that. And now let us hear a true account of the next act. How came the
stone into the goose, and how came the goose into the open market? Tell us the truth, for there lies
your only hope of safety."

Ryder passed his tongue over his parched lips. "I will tell you it just as it happened, sir," said
he. "When Horner had been arrested, it seemed to me that it would be best for me to get away with
the stone at once, for I did not know at what moment the police might not take it into their heads
to search me and my room. There was no place about the hotel where it would be safe. I went out, as
if on some commission, and I made for my sister's house. She had married a man named Oakshott, and
lived in Brixton Road, where she fattened fowls for the market. All the way there every man I met
seemed to me to be a policeman or a detective; and, for all that it was a cold night, the sweat was
pouring down my face before I came to the Brixton Road. My sister asked me what was the matter, and
why I was so pale; but I told her that I had been upset by the jewel robbery at the hotel. Then I
went into the back yard and smoked a pipe and wondered what it would be best to do.