103 Holmefield Avenue
"Well, really!" he cried, and then he choked and laughed again until he was obliged to lie back,
limp and helpless, in the chair.

"What is it?"

"It's quite too funny. I am sure you could never guess how I employed my morning, or what I ended by
doing."

"I can't imagine. I suppose that you have been watching the habits, and perhaps the house, of Miss
Irene Adler."

"Quite so; but the sequel was rather unusual. I will tell you, however. I left the house a little
after eight o'clock this morning in the character of a groom out of work. There is a wonderful
sympathy and freemasonry among horsey men. Be one of them, and you will know all that there is to
know. I soon found Briony Lodge. It is a bijou villa, with a garden at the back, but built out in
front right up to the road, two stories. Chubb lock to the door. Large sitting-room on the right
side, well furnished, with long windows almost to the floor, and those preposterous English window
fasteners which a child could open. Behind there was nothing remarkable, save that the passage
window could be reached from the top of the coach-house. I walked round it and examined it closely
from every point of view, but without noting anything else of interest.

"I then lounged down the street and found, as I expected, that there was a mews in a lane which runs
down by one wall of the garden. I lent the ostlers a hand in rubbing down their horses, and received
in exchange twopence, a glass of half and half, two fills of shag tobacco, and as much information
as I could desire about Miss Adler, to say nothing of half a dozen other people in the neighbourhood
in whom I was not in the least interested, but whose biographies I was compelled to listen to."