29 Corbet Close
"And has your business been attended to in your absence?"

"Nothing to complain of, sir. There's never very much to do of a morning."

"That will do, Mr. Wilson. I shall be happy to give you an opinion upon the subject in the course of
a day or two. To-day is Saturday, and I hope that by Monday we may come to a conclusion."