9 Harvey Street
"I did as he ordered, and when the lawyer arrived I was asked to step up to the room. The fire was
burning brightly, and in the grate there was a mass of black, fluffy ashes, as of burned paper,
while the brass box stood open and empty beside it. As I glanced at the box I noticed, with a start,
that upon the lid was printed the treble K which I had read in the morning upon the envelope.

"'I wish you, John,' said my uncle, 'to witness my will. I leave my estate, with all its advantages
and all its disadvantages, to my brother, your father, whence it will, no doubt, descend to you. If
you can enjoy it in peace, well and good! If you find you cannot, take my advice, my boy, and leave
it to your deadliest enemy. I am sorry to give you such a two-edged thing, but I can't say what turn
things are going to take. Kindly sign the paper where Mr. Fordham shows you.'