35 Glendale
"I have ordered a carriage," said Lestrade as we sat over a cup of tea. "I knew your energetic
nature, and that you would not be happy until you had been on the scene of the crime."

"It was very nice and complimentary of you," Holmes answered. "It is entirely a question of
barometric pressure."

Lestrade looked startled. "I do not quite follow," he said.