117 Beaconside
"'Tut, tut!' he cried. 'This is all quite beside the question. The point is, have you or have you
not the bearing and deportment of a lady? There it is in a nutshell. If you have not, you are not
fitted for the rearing of a child who may some day play a considerable part in the history of the
country. But if you have why, then, how could any gentleman ask you to condescend to accept anything
under the three figures? Your salary with me, madam, would commence at £100 a year.'

"You may imagine, Mr. Holmes, that to me, destitute as I was, such an offer seemed almost too good
to be true. The gentleman, however, seeing perhaps the look of incredulity upon my face, opened a
pocket-book and took out a note.